conference now made up of sessions
Instead of being a list of events, a conference is now a list of
sessions. However, an unnamed session is just a list of events, so this
is a strict generalization.

Conference: The events_to_sessions groups events into sessions. Few
other functions added to get information about a session.

Schedule tables redesigned. Sessions are indicated and various other
stylistic changes necessary as a result.