remove dependency on opam-repo-web
opam-repo-web is a small opam repository in which I had provided some
JavaScript packages. However, it is better to use the mature NPM
package repository whenever possible. Some of the packages supported in
opam-repo-web are not in NPM. In these cases, we’re wget’ing them
directly.

For some reason, NPM doesn’t support foundation 5.3.3, which is what
our scss was based on. For now, I’m going ahead with the latest version
6.2.3 (also foundation changed its name to foundation-sites). I made
just enough changes to our scss files to make them compile. However,
note the styling is completely broken in this commit, but at least
everything builds.