partially support new design on conference main pages
Update 2016’s main page to support the new design. Only partially done.
The header, new action bar, main introductory text, and a new schedule
are supported. The new profile section with listing the organizers is
not yet supported. Pages for other years not yet updated.

Replaced 2016/index.md with 2016/index.html. Conference main pages have
a lot of structure, so it easier to write the html directly.

New menu replaces the previous nav bar. It now is a child of the
header, so it can no longer be part of the general template/main.html.
Thus, the app now supports printing it, and 2016/index.html includes it
itself. The app and new Pages module support this.

Schedule is printed according to new design’s requirements. The app and
Conference.schedule function support this. The old functions are still
present but should eventually be replaced with this one.

opam: Now depend on tyxml. Trying to transition to Tyxml and possibly
the new Markup library, in place of Nethtml.