Html.map_links: use Markup lib
Implementation of this function now uses OCaml’s new Markup library
instead of Nethtml. For one thing, Markup is used by tyxml-ppx, so this
let’s us converge on these newer html libraries and eventually stop
using Nethtml.

An immediate issue was that Nethtml doesn’t retain the structure of <a>
tags with <div> childs, which the new design assumes. This is related
to an issue discussed
[here](https://sourceforge.net/p/ocamlnet/mailman/message/32318044/).