replace all 172 json files with html
All 172 json files in the repo corresponded to a specification for a
redirect. Replaced all with the corresponding html file.

Turns out there was a bug in the code before, and all the json files
under the videos/ directory were never being processed. Thus, all those
supposed links were in fact broken.