move some logic from Asset to app
The command line app now has a new sub-command `build`, which
implements the build of events, general html pages, and general
markdown pages. This logic was handled in a more automatic way in the
Asset module but that seems like overkill now. This commit moves us
toward eliminating Asset.

The Makefile now calls the app as needed on specific files.

This commit stops compiling files for 2009 - 2015. The styling is all
wrong for those years, and we’re anticipating to avoid compiling them
further. Will likely check-in the generated html and freeze those old
files.