fix menu for future years
Our menu is designed to serve also as a page title, by visualizing
saying “CUFP 2016” for instance. As it was, in 2017, when someone was
browsing a 2016 event, it would say “CUFP 2017”, which would be
confusing. Now, the topmost item, which is visible without hovering
over the menu, will say the year corresponding to the page being looked
at. For pages unrelated to the year, e.g. about.html, it will say the
latest year.